requirements.txt
anyio==4.7.0
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
et_xmlfile==2.0.0
fonttools==4.55.2
geopandas==1.0.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
kiwisolver==1.4.7
matplotlib==3.9.3
numpy==2.2.0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pyogrio==0.10.0
pyparsing==3.2.0
pyproj==3.7.0
python-dateutil==2.9.0.post0
python-telegram-bot==21.9
pytz==2024.2
requests==2.32.3
shapely==2.0.6
six==1.17.0
sniffio==1.3.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
